CXP-1291: moved useInfiniteLocales to the shared namespace
