+++ components/catalogs/front/src/components/ProductValueFilters/components/FilterLocale.tsx
@@ -4,8 +4,8 @@ import {MultiSelectInput} from 'akeneo-design-system';
 import {ProductValueFiltersValues} from '../models/ProductValueFiltersValues';
 import {useUniqueEntitiesByCode} from '../../../hooks/useUniqueEntitiesByCode';
 import {useLocalesByCodes} from '../../../hooks/useLocalesByCodes';
-import {Locale} from '../models/Locale';
-import {useInfiniteLocales} from '../hooks/useInfiniteLocales';
+import {Locale} from '../../../models/Locale';
+import {useInfiniteLocales} from '../../../hooks/useInfiniteLocales';
 
 type Props = {
     productValueFilters: ProductValueFiltersValues;
