+++ components/catalogs/front/src/hooks/useInfiniteLocales.ts
