+++ components/catalogs/front/src/models/Locale.d.ts
